Ultimo estado del proyecto

+++ src/modelo/Actor.java
@@ -23,7 +23,7 @@ public abstract class Actor {
 		return "Actor [id=" + id + ", contacto=" + contacto + "]";
 	}
 
-	protected boolean validarIdentificadorUnico(long identificador) {
+	protected boolean validarIdentificadorUnico(int identificador) {
         String s = Long.toString(identificador);
         return s.length()>8 || s.length()>11;
     }
